+++ .idea/artifacts/Toolcat_jvm_1_1_0.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="Toolcat-jvm-1.1.0">
+    <output-path>$PROJECT_DIR$/build/libs</output-path>
+    <root id="archive" name="Toolcat-jvm-1.1.0.jar">
+      <element id="module-output" name="Toolcat.jvmMain" />
+    </root>
+  </artifact>
+</component>
